refactor: move resolveManifestFile() to shared utils
