+++ packages/contentful--app-scripts/test/resolve-manifest-file/fixtures/another-contentful-app-manifest.json
